\name{fivaqks}
\alias{fivaqks}
\docType{data}
\title{
  Fiji and Vanuatu Earthquake Data
}
\description{
  Times and magnitudes (Richter scale) of earthquakes in the regions
  of Fiji and Vanuatu, for the period 1990-2015.
}

\usage{data(fivaqks)}

\format{
  A data set containing 22 variables for earthquakes around Fiji 
  and Vanuatu from 1991 to 2015.
\describe{
  \item{time}{Time of quake}
  \item{latitude}{Latitude of the quake}
  \item{longitude}{Longitude of the quake}
  \item{mag}{Magnitude of the quake}
}
}
\source{
  United States Geological Survey (USGS)
}

\examples{
  data(fivaqks)
}
\keyword{datasets}
